update_w3mmd_elo: Support Gem Elo tracking
Issues: nvs/gem#299

As W3MMD only supports player flags for a single `winner` and `loser`
(i.e. teams), a different approach must be used for Gem. Hence, the
reliance on the `ranking` variable (base-one indexing). Each game of Gem
will be considered as up to `N * (N - 1) / 2` separate matchups,
assuming that `N` players opt-in to be rated. It should be noted that
only the values specified by the `ranking` variable will affect player
ratings.

A player's `ranking` in a Gem game falls within the range `[1, N]`,
where `N` represents the number of rated players. Values outside this
range will cause the game to not be scored. Duplicates of values may
exist, so long as they fall within the specified range.

There are two cases when comparing `ranking` values in a matchup:

- Inequality: This implies that someone has won. They have a higher
  `ranking` (that is, a lower value) than their opponent. As such, ratings
  are adjusted, and both wins and loses are incremented.
- Equality: This implies that a draw occured. As such, ratings do
  not change, and neither wins or loses are incremented.

There are a few other conditions that must be satisfied, due to the
design of the script. Do note that these have no bearing upon how we
rate players, but must exist for the script to function as intended:

1. There must be at laest one player flagged as `winner` and one flagged
   as `loser`.
2. A player must be flagged as either a `winner` or a `loser` in order
   to be rated at all.

So long as these two conditions are satisfied, the script will function
as expected. For our purposes, it is recommended to have first placed be
flagged as the `winner`, and all other rated players are flagged as
`loser`. All non-rated players should be flagged as `practicing`.